move notebook updates to a separate script instead of doing it at startup

+++ fornax-base/update-notebooks.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
 # if something fails, keep going, we don't want to the stop the server from loading
-set +e
+set -e
 timeout=10
 
 # if NOTEBOOK_DIR is not defined; return or exit
-if test -z $ADD_NOTEBOOKS; then
+if test -z $NOTEBOOK_DIR; then
     # since this script is sourced, we return
-    echo "ADD_NOTEBOOKS not defined; returning ..."
+    echo "NOTEBOOK_DIR not defined; returning ..."
     return 0
 fi
 
@@ -27,32 +27,13 @@ cd $NOTEBOOK_DIR
 echo "Cloning the notebooks to $NOTEBOOK_DIR ..."
 for repo in ${notebook_repos[@]}; do
     name=`echo $repo | sed 's#.*/\([^/]*\)\.git#\1#'`
-    if ! test -f $NOUPDATE; then
-        # use nbgitpuller
-        timeout $timeout $JUPYTER_DIR/bin/gitpuller $repo main $name
-    fi
+    # use nbgitpuller
+    timeout $timeout $JUPYTER_DIR/bin/gitpuller $repo main $name
 done
 
-# change the default kernels in the notebooks;
-# remove once they are changed upstream
-if $JUPYTER_DIR/bin/jupyter kernelspec list  | grep multiband_photometry; then
-    cd $NOTEBOOK_DIR/fornax-demo-notebooks
-    jupytext --set-kernel py-multiband_photometry forced_photometry/multiband_photometry.md
-    jupytext --set-kernel py-light_curve_generator light_curves/light_curve_generator.md
-    jupytext --set-kernel py-light_curve_classifier light_curves/light_curve_classifier.md
-    jupytext --set-kernel py-ml_agnzoo light_curves/ML_AGNzoo.md
-    jupytext --set-kernel py-scale_up light_curves/scale_up.md
-    jupytext --set-kernel py-spectra_generator spectroscopy/spectra_generator.md
-fi
-# ----------- #
-
 # bring in the intro page
-if test -f $JUPYTER_DIR/introduction.html; then
+if test -f $JUPYTER_DIR/introduction.html && ! test -L $NOTEBOOK_DIR/introduction.html; then
     cp $JUPYTER_DIR/introduction.html $NOTEBOOK_DIR
-fi
-# remove any old introduction.md
-if test -f $NOTEBOOK_DIR/introduction.md; then
-    rm $NOTEBOOK_DIR/introduction.md
 fi
 
 # make symlink for NOTEBOOD_DIR in HOME
